Refactor: Reduce import lines of the dto models
