+++ src/core/auth/interface/request-user.interface.ts
@@ -1,8 +1,6 @@
 import { Request } from 'express';
 import { UserEntity } from '../../../models/entities/user.entity';
 
-interface RequestWithUser extends Request {
+export class RequestWithUser extends Request {
   user: UserEntity;
 }
-
-export default RequestWithUser;
