+++ src/core/blogs/dto/index.ts
@@ -1,0 +1,2 @@
+export { CreateBlogDto } from './create-blog.dto';
+export { UpdateBlogDto } from './update-blog.dto';
