+++ src/core/categories/dto/index.ts
@@ -1,0 +1,2 @@
+export { CreateCategoryDto } from './create-category.dto';
+export { UpdateCategoryDto } from './update-category.dto';
