+++ src/core/comments/dto/index.ts
@@ -1,0 +1,2 @@
+export { CreateCommentDto } from './create-comment.dto';
+export { UpdateCommentDto } from './update-comment.dto';
