+++ src/core/tags/dto/index.ts
@@ -1,0 +1,2 @@
+export { CreateTagDto } from './create-tag.dto';
+export { UpdateTagDto } from './update-tag.dto';
