+++ src/core/users/dto/index.ts
@@ -1,0 +1,3 @@
+export { CreateUserDto, RegisterUserDto } from './create-user.dto';
+export { UpdateUserDto } from './update-user.dto';
+export { LoginUserDto } from './login-user.dto';
